application controller spec
